Safari-611.1.10/LayoutTests/http/tests/navigation/page-cache-xhr-in-loading-iframe-expected.txt

Tests that a page with a loading iframe that has a pending XHR is able to enter the back/forward cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
* iframe starting XHR
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS XHR finished after restoring from the cache
PASS successfullyParsed is true

TEST COMPLETE
